Permit RPIO to control RPi camera LED
Permit usage of GPIO 5 (in BCM mode) which controls the camera LED for
the RPi's CSI camera module. This patch merely implements compatibility
with the RPi.GPIO package in as much as it permits control of GPIO 5 in
BCM mode (by mapping it to something other than -1) but doesn't
necessarily provide an accurate mapping (if one exists). Hence, the LED
is only controllable from BCM mode, not BOARD mode (this is equivalent
to the RPi.GPIO library).